Improved error handling
When receiving an error from openapi client, we cannot always see the exact
error reason. For example when we try to connect to a Lemmy server that no
longer exists (the domain was shut down) - the server responds with text/html
content while we expect application/json. In this case all we get is a
ClientError with underlayingError set to be RuntimeError which is an internal
type.

We shouldn't assert in these known cases.